pyproject.toml
[project]
name = "adaptive-rag"
version = "0.1.0"
dependencies = [
    "streamlit",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langchain-qdrant",
    "langchain-text-splitters",
    "langgraph",
    "langsmith",
    "qdrant-client",
    "tiktoken",
    "python-dotenv",
    "pydantic",
    "beautifulsoup4",
    "structlog",
    "ragas",
    "langgraph-cli[inmem]>=0.4.19",
    "fastembed>=0.8.0",
    "psycopg2-binary>=2.9.11",
]

[tool.setuptools.packages.find]
include = ["app*"]
